+++ frontend/src/views/sql-editor/EditorCommon/EditorAction.vue
@@ -165,7 +165,7 @@ import { useSQLEditorContext } from "../context";
 import AdminModeButton from "./AdminModeButton.vue";
 import BatchQueryDatabasesSelector from "./BatchQueryDatabasesSelector.vue";
 import DatabaseChooser from "./DatabaseChooser.vue";
-import OpenAIButton from "./OpenAIButton.vue";
+import OpenAIButton from "./OpenAIButton";
 import QueryContextSettingPopover from "./QueryContextSettingPopover.vue";
 import SchemaChooser from "./SchemaChooser.vue";
 import SharePopover from "./SharePopover.vue";
