+++ frontend/src/views/sql-editor/EditorCommon/OpenAIButton/index.ts
@@ -1,0 +1,3 @@
+import OpenAIButton from "./OpenAIButton.vue";
+
+export default OpenAIButton;
